+++ app/src/main/AndroidManifest.xml
@@ -28,15 +28,15 @@
             android:label="@string/label_settings" />
 
         <activity
-            android:name=".activity.bottom_app_bar.BottomAppBarTypeActivity"
+            android:name="com.numero.material_gallery.components.appbar.bottom.BottomAppBarTypeActivity"
             android:label="@string/bottom_app_bar" />
 
         <activity
-            android:name=".activity.bottom_app_bar.BottomAppBarActivity"
+            android:name="com.numero.material_gallery.components.appbar.bottom.BottomAppBarActivity"
             android:label="@string/bottom_app_bar" />
 
         <activity
-            android:name=".activity.bottom_app_bar.HideOnScrollActivity"
+            android:name="com.numero.material_gallery.components.appbar.bottom.HideOnScrollActivity"
             android:label="@string/hide_on_scroll" />
 
         <activity
@@ -124,24 +124,24 @@
             android:label="@string/text_field" />
 
         <activity
-            android:name=".activity.top_app_bar.TopAppBarTypeActivity"
+            android:name="com.numero.material_gallery.components.appbar.top.TopAppBarTypeActivity"
             android:label="@string/top_app_bar" />
 
         <activity
-            android:name=".activity.top_app_bar.ActionBarActivity"
+            android:name="com.numero.material_gallery.components.appbar.top.ActionBarActivity"
             android:label="@string/actionbar" />
 
         <activity
-            android:name=".activity.top_app_bar.ToolbarActivity"
+            android:name="com.numero.material_gallery.components.appbar.top.ToolbarActivity"
             android:label="@string/toolbar" />
 
         <activity
-            android:name=".activity.top_app_bar.LiftOnScrollActivity"
+            android:name="com.numero.material_gallery.components.appbar.top.LiftOnScrollActivity"
             android:label="@string/lift_on_scroll"
             android:theme="@style/Theme.MaterialGallery.DayNight.SurfaceStatus" />
 
         <activity
-            android:name=".activity.top_app_bar.CollapsingActivity"
+            android:name="com.numero.material_gallery.components.appbar.top.CollapsingActivity"
             android:label="@string/collapsing" />
 
         <activity
