+++ app/src/main/java/com/numero/material_gallery/components/appbar/bottom/BottomAppBarType.kt
@@ -1,11 +1,12 @@
-package com.numero.material_gallery.model
+package com.numero.material_gallery.components.appbar.bottom
 
 import androidx.annotation.StringRes
 import com.numero.material_gallery.R
+import com.numero.material_gallery.model.ListItem
 
 enum class BottomAppBarType(
         @StringRes override val titleRes: Int
-):ListItem {
+) : ListItem {
     BOTTOM_APP_BAR(
             R.string.bottom_app_bar
     ),
