+++ app/src/main/java/com/numero/material_gallery/components/appbar/bottom/HideOnScrollActivity.kt
@@ -1,4 +1,4 @@
-package com.numero.material_gallery.activity.bottom_app_bar
+package com.numero.material_gallery.components.appbar.bottom
 
 import android.content.Context
 import android.content.Intent
@@ -7,14 +7,13 @@ import android.view.Menu
 import android.view.MenuItem
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.google.android.material.bottomappbar.BottomAppBar
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
 import com.numero.material_gallery.repository.ConfigRepository
 import kotlinx.android.synthetic.main.activity_bottom_app_bar.*
 import org.koin.android.ext.android.inject
 
-class BottomAppBarActivity : AppCompatActivity(R.layout.activity_bottom_app_bar) {
+class HideOnScrollActivity : AppCompatActivity(R.layout.activity_hide_on_scroll) {
 
     private val configRepository by inject<ConfigRepository>()
 
@@ -28,28 +27,7 @@ class BottomAppBarActivity : AppCompatActivity(R.layout.activity_bottom_app_bar)
         }
 
         fab.setOnClickListener {
-            Toast.makeText(this@BottomAppBarActivity, "Clicked FAB", Toast.LENGTH_SHORT).show()
-        }
-
-        fabPositionRadioGroup.setOnCheckedChangeListener { _, id ->
-            when (id) {
-                R.id.attachedCenterRadioButton -> bottomAppBar.fabAlignmentMode = BottomAppBar.FAB_ALIGNMENT_MODE_CENTER
-                R.id.attachedEndRadioButton -> bottomAppBar.fabAlignmentMode = BottomAppBar.FAB_ALIGNMENT_MODE_END
-            }
-        }
-
-        fabAnimationRadioGroup.setOnCheckedChangeListener { _, id ->
-            when (id) {
-                R.id.scaleRadioButton -> bottomAppBar.fabAnimationMode = BottomAppBar.FAB_ANIMATION_MODE_SCALE
-                R.id.slideRadioButton -> bottomAppBar.fabAnimationMode = BottomAppBar.FAB_ANIMATION_MODE_SLIDE
-            }
-        }
-
-        fabVisiblyRadioGroup.setOnCheckedChangeListener { _, id ->
-            when (id) {
-                R.id.showRadioButton -> fab.show()
-                R.id.hideRadioButton -> fab.hide()
-            }
+            Toast.makeText(this@HideOnScrollActivity, "Clicked FAB", Toast.LENGTH_SHORT).show()
         }
 
         bottomAppBar.replaceMenu(R.menu.bottom_app_bar)
@@ -75,6 +53,6 @@ class BottomAppBarActivity : AppCompatActivity(R.layout.activity_bottom_app_bar)
     }
 
     companion object {
-        fun createIntent(context: Context): Intent = Intent(context, BottomAppBarActivity::class.java)
+        fun createIntent(context: Context): Intent = Intent(context, HideOnScrollActivity::class.java)
     }
 }
