+++ app/src/main/java/com/numero/material_gallery/components/appbar/top/ToolbarActivity.kt
@@ -1,24 +1,18 @@
-package com.numero.material_gallery.activity.bottom_app_bar
+package com.numero.material_gallery.components.appbar.top
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
-import com.numero.material_gallery.repository.ConfigRepository
-import kotlinx.android.synthetic.main.activity_bottom_app_bar.*
-import org.koin.android.ext.android.inject
+import kotlinx.android.synthetic.main.activity_toolbar.*
 
-class HideOnScrollActivity : AppCompatActivity(R.layout.activity_hide_on_scroll) {
-
-    private val configRepository by inject<ConfigRepository>()
+class ToolbarActivity : AppCompatActivity(R.layout.activity_toolbar) {
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        setTheme(configRepository.themeRes)
         super.onCreate(savedInstanceState)
         setSupportActionBar(toolbar)
 
@@ -26,11 +20,7 @@ class HideOnScrollActivity : AppCompatActivity(R.layout.activity_hide_on_scroll)
             setDisplayHomeAsUpEnabled(true)
         }
 
-        fab.setOnClickListener {
-            Toast.makeText(this@HideOnScrollActivity, "Clicked FAB", Toast.LENGTH_SHORT).show()
-        }
-
-        bottomAppBar.replaceMenu(R.menu.bottom_app_bar)
+        initViews()
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -45,14 +35,22 @@ class HideOnScrollActivity : AppCompatActivity(R.layout.activity_hide_on_scroll)
                 true
             }
             android.R.id.home -> {
-                finish()
+                onBackPressed()
                 true
             }
             else -> super.onOptionsItemSelected(item)
         }
     }
 
-    companion object {
-        fun createIntent(context: Context): Intent = Intent(context, HideOnScrollActivity::class.java)
+    private fun initViews() {
+        val toolbarList = listOf(primaryToolbar, surfaceToolbar, primarySurfaceToolbar)
+        toolbarList.forEach {
+            it.inflateMenu(R.menu.menu_action_bar)
+        }
     }
+
+    companion object {
+        fun createIntent(context: Context): Intent = Intent(context, ToolbarActivity::class.java)
+    }
+
 }
