+++ app/src/main/java/com/numero/material_gallery/components/appbar/top/TopAppBarType.kt
@@ -1,7 +1,8 @@
-package com.numero.material_gallery.model
+package com.numero.material_gallery.components.appbar.top
 
 import androidx.annotation.StringRes
 import com.numero.material_gallery.R
+import com.numero.material_gallery.model.ListItem
 
 enum class TopAppBarType(
         @StringRes override val titleRes: Int
