+++ app/src/main/java/com/numero/material_gallery/components/appbar/top/TopAppBarTypeActivity.kt
@@ -1,4 +1,4 @@
-package com.numero.material_gallery.activity.top_app_bar
+package com.numero.material_gallery.components.appbar.top
 
 import android.content.Context
 import android.content.Intent
@@ -8,7 +8,6 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.numero.material_gallery.R
-import com.numero.material_gallery.model.TopAppBarType
 import com.numero.material_gallery.view.ListItemAdapter
 import kotlinx.android.synthetic.main.activity_top_app_bar_type.*
