+++ app/src/main/res/layout/activity_bottom_app_bar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.bottom_app_bar.BottomAppBarActivity">
+    tools:context=".components.appbar.bottom.BottomAppBarActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
