+++ app/src/main/res/layout/activity_collapsing.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".activity.top_app_bar.CollapsingActivity">
+    tools:context=".components.appbar.top.CollapsingActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appbar"
