+++ app/src/main/res/layout/activity_toolbar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".activity.top_app_bar.ToolbarActivity">
+    tools:context=".components.appbar.top.ToolbarActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
