+++ app/src/main/res/layout/activity_top_app_bar_type.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.top_app_bar.TopAppBarTypeActivity">
+    tools:context=".components.appbar.top.TopAppBarTypeActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
